refactor: 거래내역 페이지에서 데이터를 받도록 수정
