+++ src/app/(main)/auction/page.tsx
@@ -2,7 +2,7 @@
 
 import CategoryLayout from "@/components/commons/FilterableListLayout";
 import AuctionList from "@/components/page/auction/List";
-import { mockItems } from "@/data/mock-items";
+import { mockItems } from "@/data/mock-data";
 
 export default function Page() {
   return (
