+++ src/app/(main)/trade-log/page.tsx
@@ -2,11 +2,12 @@
 
 import CategoryLayout from "@/components/commons/FilterableListLayout";
 import TradeLogList from "@/components/page/trade-log/List";
+import { tradeLogs } from "@/data/mock-data";
 
 export default function TradeLogPage() {
   return (
     <CategoryLayout categoryStorageKey="lastSelectedCategoryTradeLog">
-      <TradeLogList />
+      <TradeLogList logs={tradeLogs} />
     </CategoryLayout>
   );
 }
